// src/types/ws-general.ts
export type TransactionInfo = {
  first_trade: boolean;
  is_sniper: boolean;
  is_developer: boolean;
  is_insider: boolean;
  animal: "fish" | "whale" | "dolphin";
  timestamp: number;
  type: "buy" | "sell" | "remove" | "add";
  value: number;
  value_usd: number;
  token_amount: number;
  sol_amount: number;
  sol_amount_usd: number;
  price: number;
  price_usd: number;
  price_str: string;
  price_usd_str: string;
  market_cap: number;
  market_cap_usd: number;
  maker: string;
  signature: string;
  buys: number;
  sells: number;
};